enhancement: move Loupe and Loader to assets/icons
