+++ src/assets/icons/Loupe.js
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const Loupe = () => (
+	<svg width='16' height='16' xmlns='http://www.w3.org/2000/svg'>
+		<g fill='none' stroke='currentColor' strokeWidth='2'>
+			<path d='M14.5 14.5l-3.72-3.72'/>
+			<circle cx='6.67' cy='6.67' r='5.33'/>
+		</g>
+	</svg>
+)
+
+
+export default Loupe
