+++ src/pages/Details/index.js
@@ -3,7 +3,7 @@ import { useFetch } from 'react-hooks-fetch'
 import { Row, Cell } from 'griding'
 import { getTitleFromURL, getKindByURL } from 'utils/kind'
 import Container from 'assets/components/Container'
-import Loader from 'components/Loader'
+import Loader from 'assets/icons/Loader'
 import { Arrow, IMDB } from 'assets/icons'
 import Text from 'assets/components/Text'
 import Button from 'components/Button'
